requirements.txt
flask>=3.1
werkzeug>=3.1
pillow>=10.0
pywebview>=5.0
pywin32>=306
pyinstaller>=6.0
